sophia_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
azure-ai-projects==1.0.0b1
azure-identity==1.15.0
azure-search-documents==11.4.0
requests==2.31.0
python-dotenv==1.0.0
